requirements.txt
altair==4.2.2
pandas==1.4.4
Pillow==9.2.0
Pillow==9.5.0
plotly==5.13.1
plotly==5.9.0
st_btn_select==0.1.2
streamlit==1.22.0
wordcloud==1.8.2.2
